refactor socket code into lib/tran_sock.[ch]
While it's handy for the samples to reach in and re-use private libmuser code,
it's better if we can reduce the dependencies such that it only re-uses the
socket transport code.
