+++ samples/null.c
@@ -40,8 +40,9 @@
 #include <errno.h>
 #include <string.h>
 
+#include "../lib/common.h"
 #include "../lib/muser.h"
-#include "../lib/muser_priv.h"
+#include "../lib/tran_sock.h"
 
 static void
 null_log(UNUSED void *pvt, UNUSED lm_log_lvl_t lvl, char const *msg)
